DOC: Improve replace docstring
